Make inital fetch and rename projects to projectsContainer
