+++ src/components/App/App.js
@@ -1,5 +1,5 @@
 import React, { Component } from 'react';
-import Projects from '../Projects/Projects'
+import ProjectsContainer from '../ProjectsContainer/ProjectsContainer'
 import { fetchProjects } from '../fetch/fetch'
 
 class App extends Component {
@@ -12,17 +12,17 @@ class App extends Component {
   };
 
   componentDidMount(){
-    this.makeColors(5)
-    this.getProjects()
+    this.makeColors();
+    this.getProjects();
   };
 
   getProjects = () => {
-    const projects = fetchProjects()
+    const projects = fetchProjects();
   }
 
-  makeColors = (num) => {
+  makeColors = () => {
     let colors = []
-    for (let i = 0; i < num; i++) {
+    for (let i = 0; i < 5; i++) {
       let color = this.genHex()
       colors.push(color)
     }
@@ -66,7 +66,7 @@ class App extends Component {
             <i className="fas fa-lock"></i>
           </div>
         </section>
-        <button onClick={() => this.makeColors(5)} className="gen-colors">Generate Colors</button>
+        <button onClick={() => this.makeColors()} className="gen-colors">Generate Colors</button>
         <form className="color-form">
           <select>
             <option>project 1</option>
@@ -75,7 +75,7 @@ class App extends Component {
           <input type="text"/>
           <button className="save-pal">Save Palette</button>
         </form>
-        <Projects />
+        <ProjectsContainer />
       </div>
     )
   }
