+++ src/styles/index.scss
@@ -1,4 +1,4 @@
 @import "./variables.scss";
 @import "../components/App/App.scss";
-@import "../components/Projects/Projects.scss";
+@import "../components/ProjectsContainer/ProjectsContainer.scss";
 @import "../components/Project/Project.scss";
